back_end/creator_companion_agent/sub_agents/caption_agent/agent.py
from google.adk.agents import LlmAgent

from . import prompt

MODEL = "gemini-2.0-flash"

caption_agent = LlmAgent(
    model=MODEL,
    name="CaptionAgent",
    description="Generates a caption for a social media post",
    instruction=prompt.CAPTION_PROMPT,
    output_key="generated_caption"
)
