back_end/requirements.txt
fastapi
uvicorn
pymongo
celery>=5.2.0
redis
google-adk==1.0.0
python-dotenv
mcp